fix warnings on linux/x86_64
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@4076 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfmd/inode.c
@@ -2002,7 +2002,7 @@ dir_dump(gfarm_ino_t i_number)
 		    name == NULL)
 			break;
 		gflog_info("entry %s (len=%d) inum %" GFARM_PRId64,
-		    name, strlen(name), inode_get_number(entry_inode));
+		    name, (int)strlen(name), inode_get_number(entry_inode));
 		free(name);
 		if (!dir_cursor_next(dir, &cursor))
 			break;
